requirements.txt
streamlit>=1.36
yt-dlp>=2024.12.0
faster-whisper>=1.2.0
ctranslate2>=4.6.0
google-generativeai>=0.8.0
# Optional: used for rendering segments table
pandas>=2.2.0
# Pin setuptools to avoid pkg_resources deprecation warnings
setuptools<81

